refactor: move PulseInsightsApiError class back to mcpToolError.ts
