+++ src/sdks/tableau/methods/pulseMethods.ts
@@ -5,13 +5,11 @@ import z from 'zod';
 import {
   McpToolError,
   PulseDisabledError,
+  PulseInsightsApiError,
   PulseInsightsDisabledError,
   PulseNotAvailableError,
 } from '../../../errors/mcpToolError.js';
-import {
-  formatPulseInsightsApiError,
-  PulseInsightsApiError,
-} from '../../../errors/pulseInsightsApiError.js';
+import { formatPulseInsightsApiError } from '../../../errors/pulseInsightsApiError.js';
 import { AxiosRequestConfig, isAxiosError } from '../../../utils/axios.js';
 import { pulseApis } from '../apis/pulseApi.js';
 import { RestApiCredentials } from '../restApi.js';
